requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
firebase-admin==6.5.0
google-generativeai==0.5.4
pydantic==2.7.0
httpx==0.27.0
python-multipart==0.0.9
